+++ app/src/main/java/ru/androidschool/intensiv/data/model/movies/MoviesCollection.kt
@@ -3,7 +3,7 @@ package ru.androidschool.intensiv.data.model.movies
 import androidx.annotation.Keep
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import ru.androidschool.intensiv.network.deserializers.MoviesCollectionDeserializer
+import ru.androidschool.intensiv.data.network.deserializers.MoviesCollectionDeserializer
 
 @Keep
 @Serializable(with = MoviesCollectionDeserializer::class)
