+++ app/src/main/java/ru/androidschool/intensiv/data/network/KotlinXJsonFactory.kt
@@ -1,4 +1,4 @@
-package ru.androidschool.intensiv.network
+package ru.androidschool.intensiv.data.network
 
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.json.Json
